+++ NetSpeedMonitor/NetTrafficStatOC.h
@@ -8,8 +8,8 @@
 @interface NetTrafficStatOC : NSObject
 @property (nonatomic, assign) NSInteger delta_ibytes;
 @property (nonatomic, assign) NSInteger delta_obytes;
-@property (nonatomic, strong) NSNumber *ibytes_per_sec;
-@property (nonatomic, strong) NSNumber *obytes_per_sec;
+@property (nonatomic, assign) double ibytes_per_sec;
+@property (nonatomic, assign) double obytes_per_sec;
 @end