+++ NetSpeedMonitor/NetTrafficStatOC.mm
@@ -22,7 +22,7 @@
 
     netTrafficStatGenerator.update();
 
-    const NetTrafficStatMap net_traffic_stat_map =
+    const NetTrafficStatMap& net_traffic_stat_map =
     netTrafficStatGenerator.get_latest_net_traffic_stat_map();
 
     [_netTrafficStatMap removeAllObjects];
@@ -30,10 +30,8 @@
         NetTrafficStatOC *netTrafficStatsOC = [[NetTrafficStatOC alloc] init];
         netTrafficStatsOC.delta_ibytes = net_traffic_stat.delta_ibytes;
         netTrafficStatsOC.delta_obytes = net_traffic_stat.delta_obytes;
-        netTrafficStatsOC.ibytes_per_sec =
-        [NSNumber numberWithDouble:net_traffic_stat.ibytes_per_sec];
-        netTrafficStatsOC.obytes_per_sec =
-        [NSNumber numberWithDouble:net_traffic_stat.obytes_per_sec];
+        netTrafficStatsOC.ibytes_per_sec = net_traffic_stat.ibytes_per_sec;
+        netTrafficStatsOC.obytes_per_sec = net_traffic_stat.obytes_per_sec;
 
         NSString *key = [NSString stringWithCString:interface_name.c_str()
                                            encoding:NSASCIIStringEncoding];